Vuforia functionality in, broke everything else.
