+++ assets/ImageTargets/IT_about.html
@@ -1,0 +1,46 @@
+<html>
+<head>
+    <meta name = "viewport" content = "user-scalable=no">
+    <meta name="apple-mobile-web-app-capable" content="yes"/>
+    <style type="text/css">
+        html {
+            -webkit-text-size-adjust: none; /* Never autoresize text */
+            word-wrap:break-word; /* break the word in an another line if it's too long */
+        }
+    </style>
+</head>
+<body>
+
+	<div style="font-family:Helvetica; font-size:14px; line-height:18px; margin: 20px auto;">
+        <div style="padding-left:15px; padding-right:15px;">
+
+            <p>The Image Targets sample shows how to detect an image target and render a simple 3D object on top of it.</p>
+            
+            <p style="margin-bottom:3px;"><strong>Key Functionality</strong></p>
+            
+            <ul style="padding-left:16px;margin-top:0px;">
+                <li>Simultaneous detection and tracking of multiple targets</li>
+                <li>Load and activate multiple device databases</li>
+                <li>Activate Extended Tracking</li>
+                <li>Manage camera functions: flash and continuous autofocus</li>
+            </ul>
+
+            <p style="margin-bottom:3px;"><strong>Targets</strong></p>
+            <p style="margin-top:0px;"><a href="http://developer.vuforia.com/sites/default/files/sample-apps/targets/imagetargets_targets.pdf">Stones, Chips and Tarmac</a></p>
+            
+            <p style="margin-bottom:3px;"><strong>Instructions</strong></p>
+            <ul style="padding-left:16px;margin-top:0px;">
+                <li>Point camera at target to view</li>
+                <li>Single tap to focus</li>
+                <li>Double tap to access options menu</li>
+            </ul>
+            
+            <p><a href="http://developer.vuforia.com/legal/EULA">Terms of Use</a></p>
+            
+            <p>&copy; 2012-2014 Qualcomm Connected Experiences, Inc. All Rights Reserved.</p>
+
+        </div>
+	</div>
+
+</body>
+</html>
