+++ res/layout/activity_main.xml
@@ -4,6 +4,14 @@
     android:layout_height="match_parent"
     tools:context="com.android.visualmimo.MainActivity" >
 
+    <ProgressBar
+            style="@android:style/Widget.ProgressBar"
+            android:id="@+id/loading_indicator"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            android:layout_centerVertical="true" />
+    
     <FrameLayout
         android:id="@+id/cameraView"
         android:layout_width="fill_parent"
