--- res/layout/camera_overlay.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/camera_overlay_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent" >
-
-        <ProgressBar
-            style="@android:style/Widget.ProgressBar"
-            android:id="@+id/loading_indicator"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerHorizontal="true"
-            android:layout_centerVertical="true" />
-        
-</RelativeLayout>
