--- src/com/android/visualmimo/ImageTargets.java
@@ -1,497 +1,0 @@
-/*===============================================================================
-Copyright (c) 2012-2014 Qualcomm Connected Experiences, Inc. All Rights Reserved.
-
-Vuforia is a trademark of QUALCOMM Incorporated, registered in the United States 
-and other countries. Trademarks of QUALCOMM Incorporated are used with permission.
-===============================================================================*/
-
-package com.android.visualmimo;
-
-import java.util.ArrayList;
-import java.util.Vector;
-
-import android.app.Activity;
-import android.content.pm.ActivityInfo;
-import android.content.res.Configuration;
-import android.graphics.Color;
-import android.hardware.Camera;
-import android.hardware.Camera.CameraInfo;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.Handler;
-import android.util.Log;
-import android.view.GestureDetector;
-import android.view.LayoutInflater;
-import android.view.MotionEvent;
-import android.view.View;
-import android.view.ViewGroup.LayoutParams;
-import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.RelativeLayout;
-import android.widget.Switch;
-import android.widget.Toast;
-
-import com.qualcomm.vuforia.CameraDevice;
-import com.qualcomm.vuforia.DataSet;
-import com.qualcomm.vuforia.ImageTracker;
-import com.qualcomm.vuforia.State;
-import com.qualcomm.vuforia.STORAGE_TYPE;
-import com.qualcomm.vuforia.Trackable;
-import com.qualcomm.vuforia.Tracker;
-import com.qualcomm.vuforia.TrackerManager;
-import com.qualcomm.vuforia.Vuforia;
-import com.qualcomm.vuforia.samples.SampleApplication.SampleApplicationException;
-import com.qualcomm.vuforia.samples.SampleApplication.SampleApplicationSession;
-import com.qualcomm.vuforia.samples.SampleApplication.utils.LoadingDialogHandler;
-import com.qualcomm.vuforia.samples.SampleApplication.utils.SampleApplicationGLView;
-import com.android.visualmimo.R;
-import com.android.visualmimo.camera.ImageProcessing;
-import com.android.visualmimo.camera.ImageTargetRenderer;
-
-public class ImageTargets extends Activity
-{
-    private static final String LOGTAG = "ImageTargets";
-    
-    SampleApplicationSession vuforiaAppSession;
-    
-    private DataSet mCurrentDataset;
-    private int mCurrentDatasetSelectionIndex = 0;
-    private int mStartDatasetsIndex = 0;
-    private int mDatasetsNumber = 0;
-    private ArrayList<String> mDatasetStrings = new ArrayList<String>();
-    
-    // Our OpenGL view:
-    private SampleApplicationGLView mGlView;
-    
-    // Our renderer:
-    private ImageTargetRenderer mRenderer;
-    
-    private GestureDetector mGestureDetector;
-    
-    private boolean mSwitchDatasetAsap = false;
-    private boolean mFlash = false;
-    private boolean mContAutofocus = false;
-    private boolean mExtendedTracking = false;
-    
-    private View mFlashOptionView;
-    
-    private RelativeLayout mUILayout;
-    
-    public LoadingDialogHandler loadingDialogHandler = new LoadingDialogHandler(this);
-    
-    
-    // Called when the activity first starts or the user navigates back to an
-    // activity.
-    @Override
-    protected void onCreate(Bundle savedInstanceState)
-    {
-        Log.d(LOGTAG, "onCreate");
-        super.onCreate(savedInstanceState);
-        
-        vuforiaAppSession = new SampleApplicationSession(this);
-        
-        startLoadingAnimation();
-        mDatasetStrings.add("StonesAndChips.xml");
-        mDatasetStrings.add("Tarmac.xml");
-        
-        vuforiaAppSession
-            .initAR(this, ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
-        
-        mGestureDetector = new GestureDetector(this, new GestureListener());
-        
-    }
-    
-    // Process Single Tap event to trigger autofocus
-    private class GestureListener extends
-        GestureDetector.SimpleOnGestureListener
-    {
-        // Used to set autofocus one second after a manual focus is triggered
-        private final Handler autofocusHandler = new Handler();
-        
-        
-        @Override
-        public boolean onDown(MotionEvent e)
-        {
-            return true;
-        }
-        
-        
-        @Override
-        public boolean onSingleTapUp(MotionEvent e)
-        {
-            // Generates a Handler to trigger autofocus
-            // after 1 second
-            autofocusHandler.postDelayed(new Runnable()
-            {
-                public void run()
-                {
-                    boolean result = CameraDevice.getInstance().setFocusMode(
-                        CameraDevice.FOCUS_MODE.FOCUS_MODE_TRIGGERAUTO);
-                    
-                    if (!result)
-                        Log.e("SingleTapUp", "Unable to trigger focus");
-                }
-            }, 1000L);
-            
-            return true;
-        }
-    }
-    
-    // Called when the activity will start interacting with the user.
-    @Override
-    protected void onResume()
-    {
-        Log.d(LOGTAG, "onResume");
-        super.onResume();
-        
-        try
-        {
-            vuforiaAppSession.resumeAR();
-        } catch (SampleApplicationException e)
-        {
-            Log.e(LOGTAG, e.getString());
-        }
-        
-        // Resume the GL view:
-        if (mGlView != null)
-        {
-            mGlView.setVisibility(View.VISIBLE);
-            mGlView.onResume();
-        }
-        
-    }
-    
-    
-    // Callback for configuration changes the activity handles itself
-    @Override
-    public void onConfigurationChanged(Configuration config)
-    {
-        Log.d(LOGTAG, "onConfigurationChanged");
-        super.onConfigurationChanged(config);
-        
-        vuforiaAppSession.onConfigurationChanged();
-    }
-    
-    
-    // Called when the system is about to start resuming a previous activity.
-    @Override
-    protected void onPause()
-    {
-        Log.d(LOGTAG, "onPause");
-        super.onPause();
-        
-        if (mGlView != null)
-        {
-            mGlView.setVisibility(View.INVISIBLE);
-            mGlView.onPause();
-        }
-        
-        // Turn off the flash
-        if (mFlashOptionView != null && mFlash)
-        {
-            // OnCheckedChangeListener is called upon changing the checked state
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1)
-            {
-                ((Switch) mFlashOptionView).setChecked(false);
-            } else
-            {
-                ((CheckBox) mFlashOptionView).setChecked(false);
-            }
-        }
-        
-        try
-        {
-            vuforiaAppSession.pauseAR();
-        } catch (SampleApplicationException e)
-        {
-            Log.e(LOGTAG, e.getString());
-        }
-    }
-    
-    
-    // The final call you receive before your activity is destroyed.
-    @Override
-    protected void onDestroy()
-    {
-        Log.d(LOGTAG, "onDestroy");
-        super.onDestroy();
-        
-        try
-        {
-            vuforiaAppSession.stopAR();
-        } catch (SampleApplicationException e)
-        {
-            Log.e(LOGTAG, e.getString());
-        }
-        
-        System.gc();
-    }
-    
-    
-    // Initializes AR application components.
-    private void initApplicationAR()
-    {
-        // Create OpenGL ES view:
-        int depthSize = 16;
-        int stencilSize = 0;
-        boolean translucent = Vuforia.requiresAlpha();
-        
-        mGlView = new SampleApplicationGLView(this);
-        mGlView.init(translucent, depthSize, stencilSize);
-        
-        mRenderer = new ImageTargetRenderer(this, vuforiaAppSession);
-        mGlView.setRenderer(mRenderer);
-        
-    }
-    
-    
-    private void startLoadingAnimation()
-    {
-        LayoutInflater inflater = LayoutInflater.from(this);
-        mUILayout = (RelativeLayout) inflater.inflate(R.layout.camera_overlay,
-            null, false);
-        
-        mUILayout.setVisibility(View.VISIBLE);
-        mUILayout.setBackgroundColor(Color.BLACK);
-        
-        // Gets a reference to the loading dialog
-        loadingDialogHandler.mLoadingDialogContainer = mUILayout
-            .findViewById(R.id.loading_indicator);
-        
-        // Shows the loading indicator at start
-        loadingDialogHandler
-            .sendEmptyMessage(LoadingDialogHandler.SHOW_LOADING_DIALOG);
-        
-        // Adds the inflated layout to the view
-        addContentView(mUILayout, new LayoutParams(LayoutParams.MATCH_PARENT,
-            LayoutParams.MATCH_PARENT));
-    }
-    
-    
-    // Methods to load and destroy tracking data.
-    public boolean doLoadTrackersData()
-    {
-        TrackerManager tManager = TrackerManager.getInstance();
-        ImageTracker imageTracker = (ImageTracker) tManager
-            .getTracker(ImageTracker.getClassType());
-        if (imageTracker == null)
-            return false;
-        
-        if (mCurrentDataset == null)
-            mCurrentDataset = imageTracker.createDataSet();
-        
-        if (mCurrentDataset == null)
-            return false;
-        
-        if (!mCurrentDataset.load(
-            mDatasetStrings.get(mCurrentDatasetSelectionIndex),
-            STORAGE_TYPE.STORAGE_APPRESOURCE))
-            return false;
-        
-        if (!imageTracker.activateDataSet(mCurrentDataset))
-            return false;
-        
-        int numTrackables = mCurrentDataset.getNumTrackables();
-        for (int count = 0; count < numTrackables; count++)
-        {
-            Trackable trackable = mCurrentDataset.getTrackable(count);
-            if(isExtendedTrackingActive())
-            {
-                trackable.startExtendedTracking();
-            }
-            
-            String name = "Current Dataset : " + trackable.getName();
-            trackable.setUserData(name);
-            Log.d(LOGTAG, "UserData:Set the following user data "
-                + (String) trackable.getUserData());
-        }
-        
-        return true;
-    }
-    
-    
-    public boolean doUnloadTrackersData()
-    {
-        // Indicate if the trackers were unloaded correctly
-        boolean result = true;
-        
-        TrackerManager tManager = TrackerManager.getInstance();
-        ImageTracker imageTracker = (ImageTracker) tManager
-            .getTracker(ImageTracker.getClassType());
-        if (imageTracker == null)
-            return false;
-        
-        if (mCurrentDataset != null && mCurrentDataset.isActive())
-        {
-            if (imageTracker.getActiveDataSet().equals(mCurrentDataset)
-                && !imageTracker.deactivateDataSet(mCurrentDataset))
-            {
-                result = false;
-            } else if (!imageTracker.destroyDataSet(mCurrentDataset))
-            {
-                result = false;
-            }
-            
-            mCurrentDataset = null;
-        }
-        
-        return result;
-    }
-    
-    
-    public void onInitARDone(SampleApplicationException exception)
-    {
-        
-        if (exception == null)
-        {
-            initApplicationAR();
-            
-            mRenderer.mIsActive = true;
-            
-            // Now add the GL surface view. It is important
-            // that the OpenGL ES surface view gets added
-            // BEFORE the camera is started and video
-            // background is configured.
-            addContentView(mGlView, new LayoutParams(LayoutParams.MATCH_PARENT,
-                LayoutParams.MATCH_PARENT));
-            
-            // Sets the UILayout to be drawn in front of the camera
-            mUILayout.bringToFront();
-            
-            // Sets the layout background to transparent
-            mUILayout.setBackgroundColor(Color.TRANSPARENT);
-            
-            try
-            {
-                vuforiaAppSession.startAR(CameraDevice.CAMERA.CAMERA_DEFAULT);
-            } catch (SampleApplicationException e)
-            {
-                Log.e(LOGTAG, e.getString());
-            }
-            
-            boolean result = CameraDevice.getInstance().setFocusMode(
-                CameraDevice.FOCUS_MODE.FOCUS_MODE_CONTINUOUSAUTO);
-            
-            if (result)
-                mContAutofocus = true;
-            else
-                Log.e(LOGTAG, "Unable to enable continuous autofocus");
-        } else
-        {
-            Log.e(LOGTAG, exception.getString());
-            finish();
-        }
-    }
-    
-    
-    public void onQCARUpdate(State state)
-    {
-        if (mSwitchDatasetAsap)
-        {
-            mSwitchDatasetAsap = false;
-            TrackerManager tm = TrackerManager.getInstance();
-            ImageTracker it = (ImageTracker) tm.getTracker(ImageTracker
-                .getClassType());
-            if (it == null || mCurrentDataset == null
-                || it.getActiveDataSet() == null)
-            {
-                Log.d(LOGTAG, "Failed to swap datasets");
-                return;
-            }
-            
-            doUnloadTrackersData();
-            doLoadTrackersData();
-        }
-    }
-    
-    
-    public boolean doInitTrackers()
-    {
-        // Indicate if the trackers were initialized correctly
-        boolean result = true;
-        
-        TrackerManager tManager = TrackerManager.getInstance();
-        Tracker tracker;
-        
-        // Trying to initialize the image tracker
-        tracker = tManager.initTracker(ImageTracker.getClassType());
-        if (tracker == null)
-        {
-            Log.e(
-                LOGTAG,
-                "Tracker not initialized. Tracker already initialized or the camera is already started");
-            result = false;
-        } else
-        {
-            Log.i(LOGTAG, "Tracker successfully initialized");
-        }
-        return result;
-    }
-    
-    
-    public boolean doStartTrackers()
-    {
-        // Indicate if the trackers were started correctly
-        boolean result = true;
-        
-        Tracker imageTracker = TrackerManager.getInstance().getTracker(
-            ImageTracker.getClassType());
-        if (imageTracker != null)
-            imageTracker.start();
-        
-        return result;
-    }
-    
-    
-    public boolean doStopTrackers()
-    {
-        // Indicate if the trackers were stopped correctly
-        boolean result = true;
-        
-        Tracker imageTracker = TrackerManager.getInstance().getTracker(
-            ImageTracker.getClassType());
-        if (imageTracker != null)
-            imageTracker.stop();
-        
-        return result;
-    }
-    
-    
-    public boolean doDeinitTrackers()
-    {
-        // Indicate if the trackers were deinitialized correctly
-        boolean result = true;
-        
-        TrackerManager tManager = TrackerManager.getInstance();
-        tManager.deinitTracker(ImageTracker.getClassType());
-        
-        return result;
-    }
-    
-    
-    @Override
-    public boolean onTouchEvent(MotionEvent event)
-    {
-        return mGestureDetector.onTouchEvent(event);
-    }
-    
-    
-    boolean isExtendedTrackingActive()
-    {
-        return mExtendedTracking;
-    }
-    
-    final public static int CMD_BACK = -1;
-    final public static int CMD_EXTENDED_TRACKING = 1;
-    final public static int CMD_AUTOFOCUS = 2;
-    final public static int CMD_FLASH = 3;
-    final public static int CMD_CAMERA_FRONT = 4;
-    final public static int CMD_CAMERA_REAR = 5;
-    final public static int CMD_DATASET_START_INDEX = 6;
-    
-    
-    private void showToast(String text)
-    {
-        Toast.makeText(this, text, Toast.LENGTH_SHORT).show();
-    }
-}
